+++ package.json
@@ -23,7 +23,6 @@
     "nodeTarget": "4.2.0",
     "bundleModules": true,
     "externals": [
-      "auth0-extension-hapi-tools@1.3.1",
       "auth0-extension-tools@1.5.2",
       "auth0@2.8.0",
       "bluebird@3.5.0",
@@ -77,7 +76,6 @@
     "@hapi/joi": "^17.1.1",
     "@ngrok/ngrok": "^1.4.1",
     "auth0": "2.27.1",
-    "auth0-extension-hapi-tools": "^1.3.1",
     "auth0-extension-tools": "^1.5.2",
     "boom": "3.2.2",
     "hapi-auth-jwt2": "10.5.1",
