+++ test/test_helper.js
@@ -2,7 +2,7 @@ const nconf = require('nconf');
 const path = require('path');
 const request = require('request');
 const { sign } = require('jsonwebtoken');
-const { handlers } = require('auth0-extension-hapi-tools');
+const handlerUtils = require('../lib/handlerUtils');
 const initServer = require('../server/index');
 const config = require('../lib/config');
 
@@ -56,7 +56,7 @@ const mockHandlers = { name: 'handlers', register: async (server, options) => {
         res(fakeApiClient());
       }
     },
-    validateHookToken: handlers.validateHookToken(
+    validateHookToken: handlerUtils.validateHookToken(
       config('AUTH0_DOMAIN'),
       config('WT_URL'),
       config('EXTENSION_SECRET')
